Rename PossibleNextActions.tsx
