+++ src/components/Task/Modal/Form/LogForm/TaskModalLogForm.tsx
@@ -56,7 +56,6 @@ import {
   TagsAutocomplete,
 } from '../Inputs/TagsAutocomplete/TagsAutocomplete';
 import { TaskPhaseAutocomplete } from '../Inputs/TaskPhaseAutocomplete/TaskPhaseAutocomplete';
-import { possibleNextActions } from '../PossibleNextActions';
 import {
   useCreateTasksMutation,
   useUpdateContactStatusMutation,
@@ -67,6 +66,7 @@ import {
   handleTaskActionChange,
   handleTaskPhaseChange,
 } from '../TaskModalHelper';
+import { possibleNextActions } from '../possibleNextActions';
 import { possiblePartnerStatus } from '../possiblePartnerStatus';
 import { possibleResults } from '../possibleResults';
